Change ffmpeg instructions.
The (now defunct) ffmpeg library url has been removed and some scripts
added that will automatically build an h264-only version of ffmpeg.

The .ps1 script makes a temporary local cygwin install with all required
dependences, and then uses it to run the .sh script, which fetches the
x264 and ffmpeg source, and builds from it. The ps1 script installs the
exe and licenses afterwards, and removes all the build data.

.gitattributes is also amended so as to preserve the UNIX line endings
for the .sh script.